[SecAudit] Removed decimals vs. DECIMALS ambiguity
DECIMALS was used as a constant to specify the precision when
constructing the superclass DetailedERC20.  The confusion is due
to the fact that the detailed specification uses the decimals
property and the DECIMALS constant was used to configure the
property.  Both the property and the configuration constant had
the same name in different casings which appeared to be a
shadow declaration but in reality wasn't.